+++ tests/Unit/WsdlManagerTest.php
@@ -6,22 +6,12 @@ use Mockery;
 use PHPUnit_Framework_TestCase;
 use Skautis\Config;
 use Skautis\Wsdl\WebServiceFactoryInterface;
+use Skautis\Wsdl\WebServiceName;
 use Skautis\Wsdl\WsdlManager;
 use Skautis\Wsdl\WebServiceInterface;
 
 class WsdlManagerTest extends PHPUnit_Framework_TestCase
 {
-
-    public function testGetSupportedWebServices(): void
-    {
-        /** @var WebServiceFactoryInterface */
-        $factory = Mockery::mock(WebServiceFactoryInterface::class);
-        $config = new Config('asd');
-        $manager = new WsdlManager($factory, $config);
-
-        $services = $manager->getSupportedWebServices();
-        $this->assertTrue(count($services) > 0);
-    }
 
     public function testGetWebService(): void
     {
@@ -36,8 +26,10 @@ class WsdlManagerTest extends PHPUnit_Framework_TestCase
         $manager = new WsdlManager($factory, $config);
 
         $this->assertNotSame($wsA, $wsB);
-        $this->assertSame($wsA, $manager->getWebService('UserManagement'));
-        $this->assertSame($wsA, $manager->getWebService('UserManagement'));
-        $this->assertSame($wsB, $manager->getWebService('ApplicationManagement'));
+        $this->assertSame($wsA, $manager->getWebService(WebServiceName::USER_MANAGEMENT));
+        $this->assertSame($wsA, $manager->getWebService(WebServiceName::USER_MANAGEMENT));
+        $this->assertSame($wsA, $manager->getWebService(WebServiceName::USER_MANAGEMENT));
+        $this->assertSame($wsB, $manager->getWebService(WebServiceName::APPLICATION_MANAGEMENT));
     }
 }
+